refactor: achata estrutura movendo o projeto para a raiz do repositorio
O repositorio estava aninhado (demo-park-api/demo-park-api), com o projeto
Spring Boot um nivel abaixo da raiz. Move pom.xml, src, .mvn, mvnw, mvnw.cmd
e .gitattributes para a raiz (renames, historico preservado) e mescla o
.gitignore do subprojeto com o da raiz. Remove o diretorio intermediario.

+++ src/main/java/com/micael/demo_park_api/dto/clienteDTO/ClientePageAbleDTO.java
@@ -22,6 +22,6 @@ public record ClientePageAbleDTO(
 
     int totalPages,
 
-    int totalElements
+    long totalElements
 
 ){}
